Rename API client
